+++ app/src/main/java/dev/nekorush14/kutimu/ui/screen/HomeScreen.kt
@@ -14,12 +14,7 @@ import androidx.compose.foundation.layout.wrapContentSize
 import androidx.compose.foundation.rememberScrollState
 import androidx.compose.foundation.verticalScroll
 import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.CheckCircle
 import androidx.compose.material.icons.filled.Inbox
-import androidx.compose.material.icons.outlined.CheckCircle
-import androidx.compose.material3.Card
-import androidx.compose.material3.CardColors
-import androidx.compose.material3.CardDefaults
 import androidx.compose.material3.Checkbox
 import androidx.compose.material3.HorizontalDivider
 import androidx.compose.material3.Icon
@@ -44,6 +39,7 @@ import dev.nekorush14.kutimu.data.Task
 import dev.nekorush14.kutimu.data.TaskCategory
 import dev.nekorush14.kutimu.data.TaskFrequency
 import dev.nekorush14.kutimu.data.local.DataSource
+import dev.nekorush14.kutimu.ui.component.HabitCard
 import dev.nekorush14.kutimu.ui.theme.KutimuTheme
 
 /**
@@ -180,79 +176,6 @@ fun HomeScreen(
 }
 
 /**
- * Habit card composable.
- *
- * To grans a habit, card show [Habit.categoryIcon], [Habit.title],
- * and completion status based on [Habit.frequency].
- *
- * @param habit Habit to be displayed.
- * @param frequencyMaxCount Maximum count of status icons to display.
- * @param modifier Modifier to be applied to the card.
- */
-@Composable
-fun HabitCard(
-    habit: Habit,
-    frequencyMaxCount: Int = 7,
-    modifier: Modifier = Modifier
-) {
-    Card(
-        colors = CardColors(
-            containerColor = MaterialTheme.colorScheme.primaryContainer,
-            contentColor = MaterialTheme.colorScheme.onSurfaceVariant,
-            disabledContainerColor = MaterialTheme.colorScheme.surfaceVariant,
-            disabledContentColor = MaterialTheme.colorScheme.onSurfaceVariant
-        ),
-        elevation = CardDefaults.cardElevation(
-            defaultElevation = dimensionResource(R.dimen.card_elevation_default)
-        ),
-        modifier = modifier
-            .fillMaxWidth()
-            .padding(dimensionResource(R.dimen.card_padding_small))
-    ) {
-        Row(
-            modifier = Modifier
-                .fillMaxWidth()
-                .padding(dimensionResource(R.dimen.card_padding_medium))
-        ) {
-            Icon(
-                imageVector = habit.categoryIcon,
-                contentDescription = null,
-                modifier = Modifier
-                    .size(dimensionResource(R.dimen.icon_size_medium))
-                    .padding(end = dimensionResource(R.dimen.card_padding_small))
-            )
-            Column {
-                Text(
-                    text = habit.title,
-                    style = TextStyle(
-                        fontWeight = MaterialTheme.typography.titleMedium.fontWeight,
-                        fontSize = MaterialTheme.typography.titleMedium.fontSize,
-                    )
-                )
-                Row(
-                    modifier = Modifier
-                        .fillMaxWidth()
-                        .padding(top = dimensionResource(R.dimen.card_padding_small))
-                ) {
-                    repeat(habit.completeCount) {
-                        Icon(
-                            imageVector = Icons.Filled.CheckCircle,
-                            contentDescription = null,
-                        )
-                    }
-                    repeat(frequencyMaxCount - habit.completeCount) {
-                        Icon(
-                            imageVector = Icons.Outlined.CheckCircle,
-                            contentDescription = null,
-                        )
-                    }
-                }
-            }
-        }
-    }
-}
-
-/**
  * Task list item composable.
  *
  * To display task, each item show [Task.categoryIcon], [Task.title],
@@ -353,25 +276,6 @@ private fun HomeScreenWithoutHabitAndTaskPreview() {
             habits = emptyList(),
             tasks = emptyList(),
             greetingMessage = "Good morning!"
-        )
-    }
-}
-
-@Preview(showBackground = true)
-@Composable
-private fun HabitCardPreview() {
-    KutimuTheme {
-        HabitCard(
-            habit = Habit(
-                id = 1,
-                title = "Title",
-                description = "Description",
-                categoryIcon = Icons.Filled.Inbox,
-                completeCount = 3,
-                frequency = HabitFrequency.Weekly,
-                category = HabitCategory.Inbox
-            ),
-            frequencyMaxCount = 7
         )
     }
 }
